requirements.txt
fastapi
uvicorn
openai
python-dotenv
spacy
azure-storage-blob
azure-ai-formrecognizer
azure-search-documents
